Change type of index to int (instead of uint8_t) (#445)
This fixes an issue with adding more than 255 Characteristics (or Descriptors).

+++ src/NimBLEService.cpp
@@ -161,7 +161,7 @@ bool NimBLEService::start() {
             // of the characteristics for the service. We create 1 extra and set it to null
             // for this purpose.
             pChr_a = new ble_gatt_chr_def[numChrs + 1];
-            uint8_t i = 0;
+            int i = 0;
             for(auto chr_it = m_chrVec.begin(); chr_it != m_chrVec.end(); ++chr_it) {
                 if((*chr_it)->m_removed > 0) {
                     continue;
@@ -190,7 +190,7 @@ bool NimBLEService::start() {
                 } else {
                     // Must have last descriptor uuid = 0 so we have to create 1 extra
                     pDsc_a = new ble_gatt_dsc_def[numDscs+1];
-                    uint8_t d = 0;
+                    int d = 0;
                     for(auto dsc_it = (*chr_it)->m_dscVec.begin(); dsc_it != (*chr_it)->m_dscVec.end(); ++dsc_it ) {
                         if((*dsc_it)->m_removed > 0) {
                             continue;
